Spike to Fix ActiveRecord Performance Problems
In attempting to debug our slow ActiveRecord queries I discovered that we are using an older version of FreeTDS on Heroku than in development. This might be the source of our slow queries. This update bumps the tiny_tds gem to make sure we have compatibility with the new buildpack. It also adds dotenv-rails for local development.